Работаю над refreshToken, нужны правки в сервисе - поиск токена по хэшу

+++ src/main/java/org/musicservice/demo/service/security/RefreshTokenService.java
@@ -29,16 +29,22 @@ public class RefreshTokenService {
     private final RefreshTokenRepository refreshTokenRepository;
     private final CookieManager cookieManager;
     private final RefreshTokenProperties refreshTokenProperties;
+    private final RefreshTokenUtil refreshTokenUtil;
 
     @Autowired
-    public RefreshTokenService(RefreshTokenRepository refreshTokenRepository, CookieManager cookieManager, RefreshTokenProperties refreshTokenProperties) {
+    public RefreshTokenService(RefreshTokenRepository refreshTokenRepository, CookieManager cookieManager, RefreshTokenProperties refreshTokenProperties, RefreshTokenUtil refreshTokenUtil) {
         this.refreshTokenRepository = refreshTokenRepository;
         this.cookieManager = cookieManager;
         this.refreshTokenProperties = refreshTokenProperties;
+        this.refreshTokenUtil = refreshTokenUtil;
     }
 
-    public RefreshToken searchByTokenHash(String hash){
-        return refreshTokenRepository.findByTokenHash(hash).orElseThrow(()-> new RefreshTokenNotFoundException("refreshToken is not a found"));
+    public String getHash(String token){
+        return refreshTokenUtil.hash(token);
+    }
+
+    public RefreshToken searchByTokenHash(String hash) {
+        return refreshTokenRepository.findByTokenHash(hash).orElseThrow(() -> new RefreshTokenNotFoundException("refreshToken is not a found"));
     }
 
     public Optional<RefreshToken> getOptTokenByHash(String hash){
@@ -48,7 +54,7 @@ public class RefreshTokenService {
     @Transactional
     public void createRefreshToken(HttpServletResponse response, User user){
         String generatedRefreshToken = RefreshTokenUtil.generateRefreshToken();
-        String hash = RefreshTokenUtil.hash(generatedRefreshToken);
+        String hash = refreshTokenUtil.hash(generatedRefreshToken);
         cookieManager.setCookie(response, generatedRefreshToken);
 
         RefreshToken refreshToken = new RefreshToken();
@@ -64,9 +70,11 @@ public class RefreshTokenService {
     @Transactional
     public void delete(HttpServletRequest request, HttpServletResponse response){
         String refreshTokenByCookie = CookieUtil.getRefreshTokenByCookie(request);
-        RefreshToken foundToken = searchByTokenHash(RefreshTokenUtil.hash(refreshTokenByCookie));
-        foundToken.getUser().setRefreshToken(null);
-        refreshTokenRepository.delete(foundToken);
+        if(refreshTokenByCookie!=null){
+            RefreshToken foundToken = searchByTokenHash(refreshTokenUtil.hash(refreshTokenByCookie));
+            foundToken.getUser().setRefreshToken(null);
+            refreshTokenRepository.delete(foundToken);
+        }
         cookieManager.clearCookie(response);
     }
